s64 stuff

+++ base.h
@@ -66,7 +66,10 @@ using s16  = int16_t;
 using u16  = uint16_t;
 using s32  = int32_t;
 using u32  = uint32_t;
-using s64  = int64_t;
+// int64_t is long on linux, but long long on windows.
+// if we use int64_t for s64 we can't simply use %lld in printf for both os's.
+// So we hardcode s64 as long long.
+using s64  = long long;
 using u64  = uint64_t;
 using f32  = float;
 using f64  = double;
